add bench
The bench output format is less offensive than hyperfine.  The latter is
already available in MacPorts.

Regrettably, the sheer amount of cpu and wallclock time (sixteen and a
half minutes!) burnt on a ghc build of this project is perhaps
disproportional to the value it provides.  A simple C implementation
would better suit my needs.  The search continues.

port -t (trace mode debug) failed on this port:

    sh: line 1: 40402 Killed: 9               /opt/local/bin/stack setup --with-gcc /usr/bin/clang --allow-different-user
    Command failed:  cd "/opt/local/var/macports/build/_Users_saj_ports_sysutils_bench/bench/work/bench-1.0.9" && /opt/local/bin/stack setup --with-gcc /usr/bin/clang --allow-different-user
    Exit code: 137

I have no idea why.  Bear that in mind if submitting this upstream.